MB-Lab: Added filtering for assets
When trying to fit an asset, if there is a lot of meshes in your
file, for example a lot of meshes for the rig shapes, it becomes
very slow going through and creating a list of all the meshes
which you can fit to the character.

Added an Exclusion field to specify meshes which you don't want
to fit. The field takes a comma or space separated patterns. Any
object name which matches any of the specified patterns will
be excluded from the list.

Signed-off-by: Amir Shehata <amir.shehata@gmail.com>